Improve documentation and organize code packages
